rename registrationOptions service to usersToRegistrationOptions as it only handles the pivot table
