+++ src/types/registration.ts
@@ -3,7 +3,9 @@ import { registrations } from '../db/registrations';
 import { z } from 'zod';
 
 const groupIds = z.string().array();
+const registrationOptionIds = z.string().array();
 
 export const insertRegistrationSchema = createInsertSchema(registrations).extend({
   groupIds,
+  registrationOptionIds,
 });
